refactor: move events related component into feature folder
