+++ app/frontend/src/features/communities/events/EventCard.tsx
