+++ app/frontend/src/features/communities/events/index.ts
@@ -1,0 +1,1 @@
+export { default as EventsSection } from "./EventsSection";
